+++ packages/code-du-travail-api/src/routes/type.ts
@@ -1,0 +1,11 @@
+// Complete definition of the Search response
+export type SearchResponse<T> = {
+  hits: {
+    total: {
+      value: number;
+    };
+    hits: {
+      _source: T;
+    }[];
+  };
+};