+++ packages/code-du-travail-frontend/src/lib/matomo/types.ts
@@ -3,6 +3,16 @@ export enum MatomoBaseEvent {
   OUTIL = "outil",
 }
 
+export enum MatomoSearchAgreementCategory {
+  AGREEMENT_SEARCH_HELP = "cc_search_help",
+  ENTERPRISE_SELECT = "enterprise_select",
+  AGREEMENT_SELECT_P1 = "cc_select_p1",
+  AGREEMENT_SELECT_P2 = "cc_select_p2",
+  ENTERPRISE_SEARCH = "enterprise_search",
+  AGREEMENT_SEARCH = "cc_search",
+  AGREEMENT_SEARCH_TYPE_OF_USERS = "cc_search_type_of_users",
+}
+
 export enum MatomoSimulatorEvent {
   CLICK_CALCUL_DETAIL = "click_calcul_detail",
   CLICK_HELP_ANCIENNETE = "click_help_anciennete",
@@ -20,6 +30,7 @@ export enum MatomoSimulatorEvent {
   SELECT_CLASSE = "select_value_classe",
   SELECT_POSITION = "select_value_position",
   SELECT_COEFFICIENT = "select_value_coefficient",
+  CLICK_PRINT = "click_print",
 }
 
 export enum MatomoAgreementEvent {