+++ packages/code-du-travail-frontend/src/outils/DureePreavisDemission/steps/InfosCCn.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-import { StepInfoCCnOptionnal } from "../../common/InfosCCn";
-
-const StepInfoCCn = (props): JSX.Element => <StepInfoCCnOptionnal {...props} />;
-
-export { StepInfoCCn };