+++ packages/code-du-travail-frontend/src/outils/common/utils/math.ts
@@ -1,0 +1,4 @@
+export const sum = (arr) => arr.reduce((sum, c) => sum + parseFloat(c), 0);
+export const round = (fl: number) => parseInt((fl * 100).toString(), 10) / 100;
+export const isNotNearZero = (toTest: string | number) =>
+  parseInt(toTest.toString(), 10) > 0;